requirements.txt
requests==2.18.4
setuptools==28.8.0
PyMySQL==0.8.0
python_scrapyd_api==2.0.1
selenium==3.11.0
Django==2.0.2
pytz==2018.3
Scrapy==1.5.0
beautifulsoup4==4.6.0
djangorestframework==3.8.2
coreapi==2.3.3
django-redis==4.8.0
msgpack==0.5.6
Gunicorn==19.8.1
redis==2.10.6
Celery==4.1.0
django-celery-results==1.0.1
django-celery-beat==1.1.1
